backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.4
django-redis==5.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-filter==24.2
drf-spectacular==0.27.2        
Pillow==10.3.0
gunicorn==22.0.0
flower==2.0.1
